requirements.txt
jieba
progressbar2